V 1.1.3 Memoized focus
