+++ package.json
@@ -27,6 +27,7 @@
     "@types/react": "^18.0.24",
     "@types/react-dom": "^18.0.8",
     "concurrently": "^7.5.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "nodemon": "^2.0.20",
     "react-scripts": "^5.0.1",
     "sqlite3": "^5.1.2",
